Step to replace concept of 'inventory' with abstraction.
To make the class more reusable.

I could not think of a better word than 'domainGSheetObject'.

It will do for now. But, if anyone can think of a better name,
I'm 'all ears'.

This class still depends on InventoryFactory so that needs attending to.

See: https://github.com/xmlsquad/gsheet-to-xml/issues/21